fiction posts display correctly now
