+++ src/app/fiction/post/page.tsx
@@ -1,12 +1,13 @@
 "use client";
 import Detail from "@/component/Detail";
+import SingleFictionPost from "@/component/SingleFictionPost";
 import { useSearchParams } from "next/navigation";
 
 export default function FictionPost() {
     const searchParams = useSearchParams();
     const isWp = searchParams.get("wp");
     if (isWp === "false") {
-        return <FictionPost />;
+        return <SingleFictionPost />;
     } else {
         return <Detail />;
     }
