--- src/component/FictionPost.tsx
@@ -1,459 +1,0 @@
-"use client";
-import styles from './styles/Detail.module.css';
-import Link from 'next/link';
-import { useSearchParams } from "next/navigation";
-
-import {posts} from '@/app/fiction/page';
-
-export default function FictionPost() {
-    const searchParams = useSearchParams();
-    const idx = searchParams.get("id");
-    if (idx != null) {
-        const i = parseInt(idx);
-        if (isNaN(i) || i < 0 || i >= posts.length) {
-            return (<>
-                <h1>Post not found</h1>
-                <Link href="/">Back to Home</Link>
-            </>);
-        }
-
-        //parallel to posts, ordered by last name, accessed by title
-        const content: {[key: string] : string[]} = {
-            "A Matter of Perspective": [
-                            "There was Something Here Once",
-
-                            "Sadie Marsh is moving to Hawaii. Everyone is talking about it. At least they are tonight. Her mom’s out for dinner. And you know what they say about when the cat’s away—the mice turn the entire house into a hotbox and get stoned in the living room with two boxes of pizza.",
-                            "Sadie changed once she entered high school. She’ll be the first person to admit to that. She got off her braces, cut and dyed her hair, and changed her wardrobe. It was all out of her own autonomous design. She grew up on those teen dramas, the romance, the parties. High school was meant to be the best years of her life. And goddammit, she was going to make sure of it. She’s hurt people’s feelings, she knows. She stopped talking to a lot of middle school friends. She heard once that it broke Lily’s heart since she “abandoned them.”",
-                            "Like hell. She’ll make new friends. They all will.",
-                            "The last person that she talked to was Alice. It’s kind of funny how that works. When they were young, Alice would follow Sadie everywhere, so long as she was wanted. Sadie often joked that it was like having her own puppy on a leash, all she’d have to do was tug and suddenly she’d be there.",
-                            "Sadie liked testing this, it was almost like a game for her. She’d see how easy it was for her to get Alice to drop everything else, and hangout with her. It was amusing, if not a little pathetic to watch the results. There was this one-time Alice was in a school play (it always was impressive how such a shy little thing could get up on stage) and she got in trouble with the director because of how many rehearsals she missed so that she could hang out with Sadie. Sadie doesn’t even remember doing anything important, they just sat together in her bedroom. Talking. Afterwards, Alice offered a small smile and said, “Thank you for inviting me over.”",
-                            "Sadie still can’t place what their relationship was. Whenever they were together, Alice was completely infatuated. But she never sought Sadie out, she’d never ask her if they could spend time together. She would always wait for Sadie to initiate.",
-                            "Still, she was a good listener. Sadie recalls. That’s how she best remembers her. A listener. She wasn’t like all the other people she spoke to. With them, they always had to add something. The last time that she spoke to Alice was because she needed a good listener.",
-                            "Almost a year ago, Sadie’s parents decided to get a divorce. God, it still makes her angry just thinking about it. She knows it was inevitable; her family has always been a pack of hungry bears. Just waiting for someone else to mess up so they can attack. But still, couldn’t they just wait until she graduated? They’d spent her whole life playing happy family, what was a few more years?",
-                            "She complained to her friends about it. And that was when they started to give all of their well-intentioned advice. And she tried. She tried it all.",
-                            "They told her to go outside. But it was the midst of fall, and she hates cold weather. She’d bundle herself up and walk to the park, only to retreat after a block or so. She’d watch the rest of the day pass from the comfort of the bedroom. The grey clouds seemed like they were mocking her.",
-                            "They said that she should distract herself. She would go out to parties with her friends, but she’d only want to talk about her parents’ divorce. Then she’d be met with annoyed but pitying faces. She’d watch her favorite shows but there was always a point in them where the main character got in a fight with a parent and she’d have to turn it off, unable to differentiate between herself and the people on screen.",
-                            "She was told that she should seek help. So, she got in touch with a therapist. She spent hours sat on a lumpy couch as a man dressed in an intentionally non-threatening sweater vest explained all the many things that were wrong with her. She only lasted a few weeks before she declared herself to be cured and canceled all future appointments.",
-                            "Some late, Tuesday night, she called Alice. She doesn’t know why she made that decision; she felt like a middle schooler again. The other girl answered. Of course, she answered. Sadie doubts she even let Alice get a word in, she just rambled. Complained about nothing and everything. The conversation must’ve lasted hours, until the sun was rising. At the end, all Alice said to her was, “Let me know how I can help you.” Unlike everyone else on the planet, she wasn’t offering advice. All she wanted was for Sadie to help her to help her. Sadie ended the phone call shortly after. They didn’t speak again after that. And likely never would again.",
-                            "There was no surprise when Alice didn’t show up to her party, she’s never been the type and it’s hardly like they are friends anymore (if they ever were in the first place). But still, she can’t help but think about her. She spent years next to that girl and she still can’t pin down who exactly she was. Their whole relationship felt like an elaborate joke that never got a punchline. Maybe someday, Sadie will ask. But for now, she’ll just enjoy her last day in Seattle before she moves.",
-                            "She’s leaving. And for all she knows, Alice doesn’t even care.",
-
-                            "Losing Face",
-                            "“Fuck.” Lee made his bed, now he has to lay in it. Currently, his bed consists of the floor of one of his friend’s cousin's trailer, alongside beer cans and trash. His whole body feels sore, aching for him to stay rooted in place and never move again. So far, that’s the plan. The crash happened only a few hours ago, he can still hear the tires screeching, ringing in his ears like church bells. He already knows he fucked up, he doesn’t need to be reminded of that tidbit. So, why won’t his mind just leave him alone? In his defense, his parents should have known better than to ask him to pick up his sister. He’s twenty, obviously he was going to be drinking. But Alice needed someone to drive her home from some school event. And he’s someone, certainly.",
-                            "Lee and his sister have never been super close. He loves her more than anything else in the world. But he can’t understand her. They don’t even bicker like other siblings do; Alice always remained insufferably calm no matter how much he pushed her. It was like talking to a wall. He just wanted to do something fun with his sister. She’s always so shy about everything. Even during the summer, he’s sure she only went out like once or twice. What kind of life is that? At least to Lee, absolutely none.",
-                            "So, yes, maybe when he spoke those fateful words, “Want to see something cool?” While they were driving home, likely already going far over the speed limit, he knew it was a bad idea. But he was just trying to show his sister some fun. So, when the car inevitably nicked the front of the car behind them as he tried to weave in-between it and the car ahead (just like you always see in the movies), which caused him to lose control of the wheel and crash through a tree, then a pole before coming to their final resting place smashed against an apartment building wall, can you really blame him? And sure—maybe jumping out of the car and running away before the police arrived, leaving his sister still stunned in the passenger seat was not the smartest decision, but he really couldn’t afford to get another DUI.",
-                            "That’s why he is here now. Laying and waiting. He doesn’t know quite for who just yet. The police? His parents? The latter blew up his phone until he had to block them. He doesn’t want to hear from them. Not now. His father is still mad at him since he learned that he isn’t going to church anymore. Lee can picture his father’s face—sweat-soaked, hard-mouthed, beat-red—as he spits words from the Holy Book. Whenever he feels that pure, blinding rage, he remembers that he is his father’s son.",
-                            "Alice reached out to him once, a single text. Mom is worried about you; you should text her if you get the chance. That was all she wrote. Lee thought that something like this would pull some reaction from her. He wonders if she was even scared when the car skidded, when the windshield shattered and fell on them like a painful rain. He didn’t see if she was bleeding. He doesn’t know if she was injured or not. If she was, she didn’t say anything. She was silent when they crashed, silent afterwards.",
-                            "Everything is different. Call him prideful, but Lee isn’t just going to go home after pulling a stunt like that. He’ll sooner hop the border (any one of them, he’s a lot of things but not picky) than face that shit. He’s never felt so embarrassed of anything in his life.",
-                            "Everything is different. And he can’t tell if Alice even cares.",
-
-                            "There is Strength in Arches",
-                            "When Lily was younger, she wanted to be an architect. She didn’t want to be a princess, or a superhero, she wanted to be the person that would craft buildings that people would study in the future. It’s funny how things change. And God, do they change quickly.",
-                            "It’s freezing on her roof, the fog clouding the moon, casting her backyard in darkness. She plops against Alice’s side; her skin is just as cold as the air. “I don’t want to grow up.”",
-                            "“I know.”", 
-                            "Alice has been her friend since kindergarten. They were opposites of each other then as they are now. Everyone said that Alice was steady. That she isn’t the person that anyone would have to worry about. With Lily, those that were nice compared her to a hare. Always hopping about, never stopping at one place at long. The less nice called her impulsive and hyper-active. She still recalls the exact day she was pulled out of class and told she needed to take “special” lessons for “special” students. One hour out of the day she’d have to go sit with a handful of other students, tucked away from everyone else. She felt so alienated. Like all her peers knew that she wasn’t as smart as they were.",
-                            "She isn’t in elementary school anymore. Now she’s an adult, she and Alice both are. And now they’re met with the terrible thought of what’s next? Alice has an idea. She always has an idea. She’s steady. She already has her plan. Lily doesn’t even know if she’ll be involved in her life once college comes around. Or if it’ll just be another Sadie incident. She clenches her fist, digging her nails into her palm. Their whole lives are changing, and Lily can’t even tell if Alice-",
-                            "“I care.”",
-                            "“Oh.” Lily pauses. She lets her fist loosen; her nails created little crescents in her skin. “You do?”",
-                            "Alice doesn’t say anything for a moment, she just watches the sky. The wind causes the trees to ruffle, a single leaf to fall to the ground. “I’m not very good at showing it.”",
-                            "“Not really.”",
-                            "A huff escapes, a mix between a laugh and a scoff. The brunette’s lips, she shifts on the roof so that she can face Lily. “I’m jealous of you.”",
-                            "Lily’s own expression quirks into a half smile, a display of disbelief. “Eh? You’re jealous of me?” Alice has always been the one out of them that has received the most praise. They had one teacher go as far as hanging up her work on the classroom wall. Lily was the type of student they sent to sit in the corner.",
-                            "“Of course, I am. You’re so...you. Shamelessly you. That’s not something I could ever match.”",
-                            "“Well, you get to be you.”",
-                            "“Sometimes.”",
-                            "Silence falls over the two friends again. Neither comfortable nor uncomfortable. Just a silence. Alice has always liked the quiet, Lily never did. So, she speaks up again: “I don’t want to grow up.” She repeats, her tone a touch softer.",
-                            "“I know.”",
-                            "“But I think we have to.”",
-                            "“Probably.”",
-                            "Lily wanted to be an architect. She thinks she bit the bullet with that dream. She never studied in high school, her grades are a solid wall of Ds and Cs. She hasn’t even started applying to colleges yet. She doesn’t even want to risk getting that rejection letter.",
-                            "“...you’ll be fine.” Alice is the one to break the silence this time, matching Lily’s gentle tone. “No matter what you do, you have people looking out for you.”",
-                            "“Do I?”",
-                            "“You have me. At the very least.”",
-                            "“You have me too.”",
-                            "Lily and Alice are different. A hare. A tortoise. And because of that difference, they have always been compared to each other, and likely always will be. As if the worth of a soul is something to be quantified. But those who look will find the truth, no matter the pace they go at, someday, they’ll both find the finish line. And they’ll do it with each other in their hearts. Because Lily cares about Alice. And Alice cares about Lily. Completely and utterly.",
-                            "“I know.”"
-                        ],
-                        "Kindly Breathe": [
-                            "T-Minus Five",
-                            "Much higher than her kid’s kite back home in Kentucky USA, and parting still, Ms. Reachers was spiraling away from the ship at the steady pace of half a mile per hour. Spiraling, drifting. floating, not falling. The spots of light through dark trail in large, slow, circular motions, and Ms. Reachers can still smell the euphoric bliss of catnip. Things and an expanse of nothing: clearer than ever they had been before. Eyes watering, she might just cry crystal flying tears. Oh, the wonder!",
-                            "“Ground control… risky… just… minutes… Yes. Over.” Said a crackling in and out voice. Ms. Reachers’ man, Thomas. His voice was messy and frayed, coming in through the sides of her head bubble. She wondered if he too had experienced the thrill of space flight. If he might hurry up and join her. She was flying. No harm in experiencing a bit of freedom. “…damn house cat.” His voice coming in again. On the faintest edge of consciousness, she could make out that annoyance and slight panic when he spoke, similar to when she brings him a present from a hunting trip. Of course, Ms. Reachers knew her man was never reasonable about the things she liked to occupy herself with. So, not letting his hissing bother her, she stretched a bit, and kept on flying. Her choice to do so. Thomas might be losing his nerve.",
-                            "Four",
-                            "Deepblue. A marble they call it. A marble is all Ms. Reachers saw when it came into view as the rotations leveled out and she slowly began facing it again. Her kid choked on one of those a couple years back, dumb little thing. The tears had overwhelmed and started flowing out her eye lids to form water bubbles in air. Jello like marbles all around her.",
-                            "Was Ms. Reachers sad? She didn’t think so. No reason to be sad. Comfort, bliss, weightlessness, and floating. Floating where? Did it matter? Not much. For the first time in a long time, she wasn’t guided by the instinctual side of her brain. That sort of thing required ground beneath her feet, and cool, fresh air. Somehow her world, or rather this reality had become so claustrophobic it forced thoughts to emerge, yet she didn’t need to think of anything at all. “I’ve got… small window… grab her.” The crackle between voice, and hopefully that’s where she’s headed.",
-                            "Three",
-	                        "Around walls sound will bounce and break, growing from concentrated pitch to a soft, large blanket in a matter of seconds. It’s not just the lack of sound, but the lack of walls for the sound to bounce off and back and forth that makes the great expanse so lifeless. Ms. Reachers could only hear her breath and occasional static from the ship accompanied by an all too familiar voice. In the silence however, she focused not just on the sounds of breath as it comes and goes, but more importantly, the volume. She sounded like she was in an empty fish tank. Breathing from the front of her face bounced right back into her ears with little loss of volume. The loudness forced attention on the movement of breath. Never before had she thought about breathing, but in these moments, it was what she could control. The only thing that responded to the amount of control she had in such aimlessness was the sounds from breathing. She hoped Thomas could hear this defiant breath of hers and understand too what it means to control air in such a way. She decided he wasn’t appreciating oxygen enough, though she didn’t fully comprehend why.",
-                            "Two",
-	                        "Dr. Thomas had quite the annoying trip so far. He had a delay making it to the ISS due to an unforeseen docking error, which wouldn’t have occurred had the program listened to his request to recalculate the mass of his cargo. He hated this cat which he was forced to bring. Ms. Reachers is what his son named her, for how much she liked to stretch and knock things over. Of course, “Ms. Reachers” hadn’t taken leaving orbit so well, which is to say she’s as scratchy and hissy as she’s ever been. She even scratched three deep scares along the length of his forearm as she held onto him for dear life when lifting off. Devil spawn, truly. Yet, he’s the one who forgot to tether her as they went for a spacewalk, just ten minutes ago. Now he had only a couple minutes before Ms. Reachers would be beyond reachable. Now back on the station, they were drifting in opposite directions and the longest tether would get him within arm’s reach right now. Despite the annoyance and obligation, he felt, he still found the sight of his high, subdued cat floating away rather amusing.",
-                            "The astronaut sighed while looking out the airlock window. “Welp, there’s no way I can ever enter my own damn house again without this stupid cat.” He mumbled to himself. Turning on his radio, he declared “This is Thomas, I’m heading out one more time, hopefully to push off in the right direction to catch the straying animal test subject. Over.” The current operator at mission control replied with a warm tone of hope and vigor, “Godspeed, Dr. Thomas, and best of luck to you and Ms. Reachers.” The operator knew the cat’s name. Thomas wondered if mission control was intercepting calls he’d had with his family, as he certainly wouldn’t call the cat that name otherwise. “Right. Ms. Reachers. Thanks.”",
-                            "One",
-	                        "The lack of oxygen likely prolonged her high, though Ms. Reachers was beginning to come down. The only problem was the drifting didn’t stop. She was starting to overthink everything now, and the spinning made her feel out of control.",
-                            "Breathe, all she needed to do was keep her breath. That’s the control In and out. Up and down. The pricks of light then, the space between. Oh, the space between! She lost breathing, it was becoming the length between breath, not breath in drawn out length. She missed the sides of concrete pathways and chasing helpless little critters. Wonderful snacks. Taste, there was no taste up here. Bland, dry food is what her man gave her. Organic lights and chaos is all she felt. She appreciated it for a moment, but considering it a little less emotionally drifting, she’d like to go home please.",
-                            "Lift-Off",
-                            "Thomas outstretched his right arm and grabbed the now wiggling cat by its right hind leg, held on by a thin cable himself. “Thank God. You’re not close to dying. Miraculously.” he said to it happier for himself than the cat. Ms. Reachers merely looked at the man and tried to nuzzle his arm before he yanked too aggressively. She reflexively scratched through her suite onto his and air started seeping out. Hopefully, if he was finally taking her back, He’d learn to control his breath too. “Godspeed Thomas.”"
-                        ],
-                        "Cruelty of Spring": [
-                            "Simone ruined my life. I never knew love before I met her, and her love has tortured me every day since.",
-                            "When I first moved to Paris, I only had a small room with a simple cot and a single east facing window that looked out onto a narrow, shaded alley. I was an aspiring painter and never imagined having passion for anything beyond that which was superficial. I had a lust for life back then, devouring all the beauty I could find and expelling it back into the world a little worse than how I found it. My move to Paris was justified by the job I lined up with a cousin working as a house painter, but in honesty, I was twenty-six and wanted to live the reality of a poor artist with the dignity of doing so in the City of Lights. I took the lush rebirth of a Parisian Spring as a sign of my own renaissance.",
-                            "Most of my days were spent painting. If it wasn’t a house, it would be a canvas. If I wasn’t painting, I would be walking the streets for inspiration. Every corner was a momentous scene, and I was never without stimuli. Vignettes of the mundane transformed into meaningful triptychs of humanity under the spell of the city. I found magic in the simple pleasures of people watching. Enjoying a jamon-beurre at a café under the sun, fragrant lavender and the blush of red clover in a park. Another thing Paris was never short in was beautiful women. Although amateur artists are commonplace here, it was easy enough to charm a woman into modelling for me, often without bribe. I’d be a fool to deny the admiration I had for these women and an even bigger fool to lie about my ulterior motives. The shallow nature of our rendezvouses would leave me satisfied, if only for an evening. Before Simone I was easily bored by any kind of emotional connection or vulnerability with the opposite sex. For me it was always a transactional affair. Before Simone, the connection I had to art and my freedom during those days of solitude was all I cared about.",
-                            "Simone seemed to just appear one night. I was at the bar across from the building I inhabited where other wayward souls would come to numb their senses with drink. The evening was like one of many before it, a cool night where the streets glimmered with the remnants of an early spring shower, carrying with it a distinct trace of petrichor. Like a ghost, she had simply materialized next to me, dark haired with delicate features that were fixed with a severe expression. I was already in a thick haze of intoxication and yet her presence dominated what little faculty remained in me. I couldn’t begin to understand the effect she had on me so immediately, but I turned to her at once with the same courage I had with every other woman I had intended to paint or bed. She was receptive but her haunting visage did not warm or soften.",
-                            "I cannot recall what transpired that night, but when I awoke the next morning, I was alone with only a broken memory and a single lucid reminder of her. The impression she left was painful with longing and lingered for days after like a bruise. I had spent the next week in a daze, hoping she would manifest before me again. Until one day she did.",
-                            "That spring was the most significant of my life. I spent all the time I could with Simone and her image emerged in every figure I painted without intention. She was luminous in the way a fire is. Full of magnetic light with a heat that emanates from a place of innate danger. We both shared a hedonistic sensibility, indulging in every amusement we could imagine. I hadn’t been sure what she did for work considering she was never without funds and her hours of freedom seemed to be nearly unrestrained. I had figured that perhaps her wealth was generational or during the hours she was not with me she held certain arrangements that allowed the maintenance of such a lifestyle. I wasn’t even sure where she stayed but either way I didn’t mind, the secret of her independence only mystified me further.",
-                            "Simone had a cruelty about her that I found at once addicting. On many occasions she would mock the beauty I assigned to the world or ridicule the meaning I placed on the art I created. The opinions of others had never bothered me, my art was by no means remarkable, but my place on this earth was mine alone to make meaningful. It was a pleasure I’d never known to be observed so frankly by this woman. Her sharp words would pierce through me, striking some hidden nerve I never knew existed. Blood would rush through me in a wave, all at once exhilarating and humiliating. She took great amusement in her malice, and it wasn’t long before the impact of her cruelty became physical, and I not only endured it but embraced it.",
-                            "One night at Simone’s suggestion, we attended a cabaret set to take place at a club in an area of town I hadn’t been familiar with. The room was saturated with excitement and seemed charged with energy so electric it clung to my skin like static. Simone met the glances of other patrons and new glasses of liquor and wine would appear upon our table without request. She would turn her head and strangers provided cigarettes without comment or question. Her silent command troubled me.",
-                            "As the night unfolded, something in Simone began to reveal itself. The natural grace in her distinguished demeanor developed an air of agitation and her black eyes, which always held such playful contempt, were veiled with something I could only describe as wicked. She caught my gaze, and her rouged lips distorted into a smile so disturbing I could’ve recoiled had I not been so paralyzed with excitement. In this moment I might’ve expected her to unleash upon me with one of her familiar cutting remarks or open-handed strikes but instead she purred, “Would you like to see my apartment?”",
-                            "She led me through the fog of perfumed smoke and beyond the cabaret doors. Her apartment was a building on the corner of the same block, discreet with an austere elegance. Her room was quiet in its decadence with a few fine pieces of polished wood furniture finished with details of pearl inlay. The west-facing window was open to a terrace and framed by white curtains with the thin, delicacy of tissue paper.",
-                            "I had been too captivated by her room to notice that Simone had gone and reappeared with two crystal glasses in hand. “To your health,” she says, bringing the lip of the glass to hers. I finished my drink at once and she removed it from my hand, placing it on the vanity. She shrugged out of her coat and pulled from the pocket a silver knife. Simone and I had been intimate plenty of times before, and yet I felt more vulnerable with her than ever in this strange place with all of my clothing on. She glided over to where I was seated on her bed and showed me this ornate object, handling it with such poise. The blade and handle were a striking etched silver and gleamed with a shine that only comes from the labor of love. She didn’t have to say what part I played in her ritual; I could already tell.",
-                            "The strange animalistic glare she held at the club had disappeared, replaced with a calmness so haunting I felt it in every hair follicle and skin cell. I had resigned myself to doing whatever she asked of me, I could never leave her. I never wanted to.",
-                            "That was the last night I saw Simone. I spent many sleepless nights and aimless days searching for her, stalking the streets and the building in which we spent our last evening. I’d manifest her presence in my hours of longing with no luck. Every night since then I would lie in bed and conjure dreams of her that morphed into nightmares. I cherish every scar she inflicted on me because it proves my love was real. Her dismissal was her final act of cruelty and humiliation. I’d never known such evil, such beauty, such excess and exquisite pain. Where I once found beauty, I found nothing that could compare to this unrequited love."
-                        ],
-                        "The Coffee Shop Encounter": [
-                            "Setting: A cozy coffee shop with warm lighting, a few tables scattered around, and a counter at the back where a barista is making drinks. It’s early evening, and the shop is half-empty.",
-                            "At Rise: JESSICA sits at a small table near the window, her laptop open, typing furiously. A cup of coffee sits untouched next to her. SAM enters, holding a book, looking around as if searching for someone. He spots Jessica and hesitates for a moment before approaching.",
-                            "SAM",
-                            "(standing near the table, awkwardly)",
-                            "Hey, uh… Jessica, right?",
-                            "JESSICA",
-                            "(not looking up)",
-                            "Depends who’s asking.",
-                            "SAM",
-                            "(chuckling nervously)",
-                            "It’s Sam. From the writer’s group? We met last week.",
-                            "JESSICA",
-                            "(looking up, recognizing him, but not entirely thrilled)",
-                            "Oh. Right. Sam.",
-                            "(She gestures to the empty chair across from her.)",
-                            "SAM",
-                            "(sitting, setting the book on the table)",
-                            "I didn’t mean to interrupt. You just looked… busy.",
-                            "JESSICA",
-                            "(flatly)",
-                            "That’s because I am.",
-                            "SAM",
-                            "(fiddling with the book)",
-                            "Got it. I’ll, uh, keep this quick. I wanted to ask about—",
-                            "JESSICA",
-                            "(interrupting, leaning back in her chair)",
-                            "Let me guess. Feedback on your chapter?",
-                            "Stage Direction: Jessica closes her laptop with a sharp snap and leans forward, folding her arms on the table.",
-                            "SAM",
-                            "(stammering)",
-                            "Well, yeah, but—it's not just that.",
-                            "JESSICA",
-                            "(smirking)",
-                            "“Not just that”? You’re off to a great start.",
-                            "SAM",
-                            "(defensively)",
-                            "Hey, it’s not easy asking for feedback from someone who clearly doesn’t want to give it.",
-                            "JESSICA",
-                            "(sighing)",
-                            "Look, Sam, I don’t mean to be harsh, but I’ve got deadlines. Multiple ones.",
-                            "SAM",
-                            "(leaning forward)",
-                            "I get that. Really, I do. But you gave great feedback last week. It helped me rethink my whole second act.",
-                            "Stage Direction: Jessica softens slightly but maintains her guarded posture.",
-                            "JESSICA",
-                            "(leaning back)",
-                            "Fine. What’s the issue?",
-                            "SAM",
-                            "(rushing)",
-                            "Okay, so you said my characters felt flat, right? I tried giving them more conflict, but now it feels… forced. Like, they’re arguing just to argue.",
-                            "JESSICA",
-                            "(muttering)",
-                            "Classic overcorrection.",
-                            "SAM",
-                            "(excitedly)",
-                            "Exactly! So what do I do? How do I make it feel real?",
-                            "Stage Direction: Jessica picks up her coffee, finally taking a sip. She considers him for a moment before responding.",
-                            "JESSICA",
-                            "(seriously)",
-                            "Conflict isn’t just about yelling or drama. It’s about tension. Subtext. What are your characters not saying?",
-                            "SAM",
-                            "(frowning)",
-                            "What they’re not saying?",
-                            "JESSICA",
-                            "(nodding)",
-                            "Yeah. Like this conversation. You think we’re talking about your story, but we’re not.",
-                            "SAM",
-                            "(confused)",
-                            "We’re not?",
-                            "JESSICA",
-                            "(flatly)",
-                            "Nope. We’re talking about your need for validation.",
-                            "SAM",
-                            "(defensive)",
-                            "That’s not—",
-                            "JESSICA",
-                            "(cutting him off)",
-                            "Relax, it’s not a bad thing. Everyone needs validation. But if you want authentic characters, you’ve got to figure out what they’re hiding.",
-                            "Stage Direction: Sam leans back, thinking this over. Jessica watches him, her expression softening just a little.",
-                            "SAM",
-                            "(quietly)",
-                            "So… what are you hiding?",
-                            "JESSICA",
-                            "(startled)",
-                            "Excuse me?",
-                            "SAM",
-                            "(sincerely)",
-                            "I mean, you clearly care about writing. About helping people get better at it. But you act like you don’t. Why?",
-                            "Stage Direction: Jessica hesitates, then lets out a bitter laugh.",
-                            "JESSICA",
-                            "(smirking)",
-                            "You really want to know?",
-                            "SAM",
-                            "(nods)",
-                            "Yeah.",
-                            "JESSICA",
-                            "(sighing)",
-                            "Fine. I care, okay? I care a lot. And sometimes it feels like I care too much. Like I’m putting all this effort into helping others while my own work sits unfinished.",
-                            "Stage Direction: Sam listens, his expression thoughtful.",
-                            "SAM",
-                            "(earnestly)",
-                            "Maybe that’s because you’re good at it. Helping others, I mean.",
-                            "JESSICA",
-                            "(bitterly)",
-                            "Yeah, well, being “good at it” doesn’t pay the bills.",
-                            "SAM",
-                            "(smiling slightly)",
-                            "Neither does writing.",
-                            "Stage Direction: Jessica laughs despite herself, the tension easing slightly.",
-                            "JESSICA",
-                            "(smirking)",
-                            "Touché.",
-                            "Stage Direction: There’s a brief silence. Sam glances at the book he brought.",
-                            "SAM",
-                            "(offering it)",
-                            "Look, if it helps, you don’t have to read the whole thing. Just the first ten pages. Tell me if it feels forced.",
-                            "JESSICA",
-                            "(rolling her eyes, but taking the book)",
-                            "Fine. Ten pages. But you owe me a coffee next week.",
-                            "SAM",
-                            "(grinning)",
-                            "Deal.",
-                            "Stage Direction: Jessica opens her laptop again as Sam gets up to leave. He pauses at the door, looking back.",
-                            "SAM",
-                            "(sincerely)",
-                            "Thanks, Jessica.",
-                            "Stage Direction: Jessica doesn’t look up, but a small smile creeps onto her face as she types.",
-                            "JESSICA",
-                            "(muttering)",
-                            "Don’t make me regret it.",
-                            "Stage Direction: Sam exits, and the sound of typing fills the coffee shop as the lights fade.",
-                            "Curtain."
-                        ],
-                        "Dr. Denise and the Case of the Town Flu": [
-                            "The sky was dark and cloudy as Denise sat in the business class of the Eurail train, watching green hills roll by. She hated traveling, yearning to retreat to her cozy apartment. But it was too late; she'd subleased it for two months.",
-                            "Just a week ago, her boss had called her in for a special assignment: a new strain of flu was draining life from a small rural town, Margottsville. Her task was to uncover its cause and find a cure.",
-                            "Margottsville wasn't the bustling city she loved. Though it had basic amenities—internet, laptops, phones, schools—its medical standards lagged. That's why she was sent: to provide what the town lacked.",
-                            "The train’s whistle jolted Denise from her thoughts; they had reached her station.",
-                            "________________________________________",
-                            "A handsome man in a police uniform greeted her.",
-                            "“Good morning, miss,” he said. “May I take your bag?”",
-                            "“Oh,” Denise replied, surprised. “Yes, please.”",
-                            "As she handed it over, he immediately dropped it, the handle slipping through his fingers. Bending down to retrieve it, his badge clattered to the ground, followed by his phone, which slipped from his pocket.",
-                            "By the time he stood up, face red and sheepish, Denise was trying not to laugh.",
-                            "“I’m sorry about that,” he stammered. “Everyone calls me ‘Slippers.’ Things just slip away from me.”",
-                            "“It’s alright,” Denise said, smiling. “Allow me?”",
-                            "He smiled back, grateful. “Thank you.”",
-                            "As she handed his things back, Denise noticed his bright green eyes and felt her face flush.",
-                            "“I’m Penbrooke. Officer Daniel Penbrooke.”",
-                            "“Pleasure,” Denise said, extending her hand. “Dr. Denise Walker.”",
-                            "“Yes, we know who you are,” Daniel said as they shook hands. “You’re a lifesaver. Literally. But we really need your help. Five people have died, two of them children… We don’t know what’s wrong, and everyone’s on edge.”",
-                            "“But you know it’s the flu, right?”",
-                            "“Actually…” Daniel hesitated. “That’s what we’ve been telling people to keep them calm. But we don’t know. The symptoms are strange. And I’m not a doctor, but I think it’s something else.”",
-                            "Denise’s heart raced. She hadn’t even finished her training yet—why did her boss think she was ready for this?",
-                            "________________________________________",
-                            "As soon as the car stopped, Daniel rushed to open Denise’s door.",
-                            "“Miss Walker,” he said, then quickly corrected himself. “I mean, Doctor! Doctor Walker! Sorry.”",
-                            "This time, Denise laughed, but not in mockery. Daniel sensed it, and his embarrassment turned into shared laughter.",
-                            "“Thank you, Daniel.” She stepped out.	",
-                            "Suddenly, a ragged old woman with silver eyes and yellow teeth charged at them. “Do not come in here! Do not come in here!”",
-                            "“Let’s go,” Daniel said, gently taking Denise’s arm and leading her toward the apartment building.",
-                            "“Do not come in here!” the woman continued to yell. “Do not drink the water! You’d rather die of thirst!”",
-                            "“Alright,” Daniel said, trying to calm her. “Have a nice day.”",
-                            "Denise kept her eyes straight ahead, avoiding the woman's furious gaze. Eventually, the woman stopped yelling and walked away.",
-                            "As they reached the lodgings, a sweet old couple greeted them.",
-                            "“A pleasure to meet you, dearie,” said the woman. “I’m Miss Lady, and I own this building. Was Daniel a gentleman?”",
-                            "“Yes,” Denise said, glancing shyly at Daniel. “He was great.”",
-                            "“Wonderful,” Miss Lady said. “This is my husband, John Monroe.”",
-                            "“They’re famous,” Daniel chimed in. “John Monroe and Miss Lady founded this town forty years ago. Back then, women couldn’t even buy a horse without their husband’s permission.”",
-                            "“Wow,” Denise said. “That’s crazy.”",
-                            "“It was a simpler time,” John Monroe said gruffly. “A better time.”",
-                            "Something in Denise jolted—something about John felt off. But she brushed it aside.",
-                            "“Hush now,” Miss Lady warned her husband with a look. “We have company.”",
-                            "John scoffed and walked away without another word.",
-                            "Miss Lady’s smile didn’t reach her eyes. “He’s fine, dear. Let me show you to your room.”",
-                            "________________________________________",
-                            "Two weeks passed, and three more people fell ill. Denise kept administering water and antibiotics, but nothing worked. She couldn’t remember her last full night of sleep.",
-                            "Daniel visited daily, always managing to make her laugh. She was grateful for his friendship.",
-                            "But one day, everything changed.",
-                            "Denise was leaving the hospital at 1:30 in the morning when someone grabbed her arm hard, pulling her back. Startled, she looked down into the eyes of the same old woman, who now gazed at her with an intensity that made Denise shake with fear.",
-                            "“Stop feeding them water,” the woman said, her voice soft for the first time. “John Monroe is poisoning the water. He’s trying to kill us all.”",
-                            "Denise would normally have shrugged her off and run away, but she remembered the haunting look in Monroe’s eyes and the nagging feeling that something was wrong.",
-                            "She called Daniel, knowing he was on the graveyard shift. They met at the hospital and spoke with the old woman.",
-                            "She was the second person to move into town after the Monroes founded it. She witnessed John Monroe’s relentless advocacy against women’s rights and his resistance to the town’s modernization.",
-                            "And she had seen him at the water conservation system that night, pouring poison into the water. She had photos on her phone, but John had threatened her life if she showed them to anyone.",
-                            "“But I don’t care anymore,” she said, her voice breaking. “My grandson died three weeks ago. He was only seven. I had to tell someone. I don’t care what John does to me.”",
-                            "The pictures were all Daniel needed to secure a warrant for Monroe’s arrest. Denise watched the events unfold with wide, frightened eyes. How could someone do this?",
-                            "The next day, the news broke: Officer Daniel “Slippers” Penbrooke had arrested the town’s founder for poisoning the water and murdering five people.",
-                            "Denise would never forget how close she came to overlooking such a crucial detail. But she was glad she had trusted her gut."
-                        ],
-                        "Two Sides of the Same Coin": [
-                            "I hadn't meant to end up at the Willow Creek Animal Shelter. My hands had simply taken the familiar turns while my mind wandered, muscle memory carrying me down the winding gravel road where I'd spent countless mornings cleaning kennels and teaching training classes. That was before – before the divorce, before Andre’s betrayal, before everything changed.",
-                            "Titan sat in the passenger seat, his graying muzzle resting on the window ledge. Technically, he was Andre’s dog now. The courts had been clear about that – property division, clean and simple. But Titan had been ours, had slept at the foot of our bed for eight years, had waited for me every morning by the coffee maker. Some bonds couldn't be severed by legal documents.",
-                            "I pulled up beside one of the storage sheds in the back, deep in thought. A few moments later, Andre’s truck pulled in behind me. I felt the familiar tightening in my chest. He wasn't supposed to be here, wasn't supposed to follow me to this sanctuary of my past. This was supposed to be my time alone with Titan - we had agreed to this arrangement. I rolled the window down, throwing a questioning look in his direction.",
-                            "\"Just looking for some wood scraps,\" he called out, as if the explanation could smooth over the intrusion. \"I’m patching up Titan’s old dog house. Denise said I could stop by anytime.\"",
-                            "Denise hadn't said that. I knew because Denise and Sarah, the shelter owners, had been there for me during the worst of it, during those months when I couldn't tell which way was up. They'd let me work with the dogs when I couldn't sleep, when the empty house echoed too loudly with Titan’s absence.",
-                            "I opened the door to face Andre, torn about how to respond to his unexpected arrival. Part of me wanted to call him out....another part was embarrassed that he had shown up here, acting as if everything between us was resolved. Before I could decide how to address the situation, Titan was out of the car and bounding across the field, his arthritic joints forgotten in his excitement. My heart lurched as he disappeared into a sea of dogs and their owners – some sort of event I hadn't known about. One moment he was there, a familiar flash of copper fur, and the next, gone. I knew I had no right to chase after him, but my feet moved anyway.",
-                            "\"Titan!\" I called, my voice carrying across the field. The name felt heavy on my tongue, weighted with everything else I'd lost. “Titan, come!” I called out to him, over and over again. He didn't come. He always used to come.",
-                            "I searched frantically. The crowd was breaking up, and I approached a blonde woman standing off to the side. I described Titan and asked if she had seen him. She looked at me solemnly. \"I know where he is,\" she said, \"I thought he was one of the shelter dogs.\" She led me to a fenced area where Titan stood chained, his tail tucked. When he saw me, he cowered, as if he'd somehow disappointed me by getting caught, by being contained. Or maybe he sensed the guilt in my own heart – for visiting him at Andre’s house when he wasn't home, for slipping him outside to walk our old neighborhood together, for refusing to make the break clean.",
-                            "\"Oh, buddy,\" I whispered, kneeling beside him. \"It's okay. You didn't do anything wrong.\" The words caught in my throat, and I realized I was talking to myself as much as to him. The woman unhooked the chain, and Titan pressed against my legs, solid and warm and real, just like he used to.",
-                            "I had no leash. No way to hold on, no way to control our path back. But as I stood there, Titan waiting patiently beside me, I understood something. Sometimes the things we think we need to hold onto – the leashes, the marriages, the perfect stories we tell ourselves – aren't what keep us safe at all. Sometimes love means letting go, even when every fiber of your being wants to hold on.",
-                            "Titan and I walked back together, unleashed but connected by something deeper than ownership. I would return him to Andre, because that was right, because that was what healing meant. Let Andre search for his wood scraps, for pieces of the past. I was finally ready to build something new, even if it meant building it alone.",
-                            "The gravel crunched beneath our feet, and Titan’s fur brushed against my leg with each step. His limping gate reminded me that nothing lasts forever, that love and loss are two sides of the same coin. But there was comfort in that truth now, where there had only been pain before. I walked on, toward Andre’s truck – not toward the home that was, but toward the understanding that sometimes we find our home in the places we least expect, and letting go was the key to moving on."
-                        ],
-                        "Perri and Jonathan the Intergalactic Door Dashers": [
-                            "Jonathan has gotten used to tuning out the stale laugh track that’s repeatedly inserted into whatever TV show from Earth that Perri is watching this time. Perri has his feet up on the control board, precariously close to buttons that navigate the ship. Despite the fact that Jonathan has warned him repeatedly, “Don’t do that. It’s against safety protocol.” To Jonathan, going against protocol is unthinkable. He struggles to comprehend how Perri can so carelessly disregard these safety measures, but is also slowly recognizing that outside of the cadet force, this is the norm.",
-                            "Focused on his push up reps, it takes a second for Jonathan to notice that Perri is calling him, “Jonathaaaaan. Jonathan! We got a gig!” He looks up and sees the blue notification bubble on the bottom right of the spacecraft screen. Finally! It felt like they have been waiting for work for centuries now. Perri convinced Jonathan to team up with him for delivery jobs when he initially left the cadet force two weeks ago. Doing delivery jobs helped keep him occupied during this change. He was still getting the hang of not having someone to boss him around all the time and the lack of instruction gave him plenty of time to contemplate what to do next with his life.",
-                            "He walks over to the spaceship screen and catches a glimpse of what he thinks is Perri’s Earth show called Seinfeld, before commanding the spacecraft’s voice assistant to bring up the details. Perri sighs at this disruption but scoots over so Jonathan can pilot the spacecraft to their destination. Jonathan reads the request as he plugs in the coordinates, “One egg from Planet 3AB-0. Should be easy enough.”",
-                            "Perri’s attention returns as he hears this, “Oh I love that planet, last time I was there I met this gorgeous-” Perri goes on about this past rendezvous while Jonathan gets the ship settings ready. He’s piloted many spacecraft before, it’s something he trained thoroughly during his time with the cadet force. But this ship, the only one that he could manage to get his hands on short notice when he made the decision to leave, was an older model that had been aggressively modded. It was just another thing that Jonathan was getting used to.",
-                            "It’s not long before Planet 3AB-0 is in sight. It’s a murky green orb with two small moons. Against the dark vastness of the rest of space, it appears to glow a bit. But as they inch closer, Jonathan realizes the green comes from the huge expanse of trees covering a majority of the planet’s surface.",
-                            "The spacecraft rumbles as it breaks through the green planet’s atmosphere. Perri and his disregard for spacecraft safety is, of course, not buckled in, so the turbulence sends him flying and screaming. Jonathan stays focused on landing, knowing that Perri is capable of handling himself despite his poor choices.",
-                            "Jonathan spots a small clearing amongst the dense foliage that makes a good landing spot and directs the ship there. As they descend, the thick swamp trees and vines block the light and it gets darker and darker. The ship sputters as the bottom of it makes contact with the mud and settles into the mushy swamp floor.",
-                            "The sound of whirring slows as the ship powers down. Perri unhinges himself from the bar that he ended up clinging to. Jonathan squints outside the ship window. As the steam clears outside, he takes a moment to take in his surroundings and evaluate what gear he might need. Jonathan is adamant about keeping his basic essentials at the ready, so it only takes him a second to pull on his waterproof boot and grab his gear pack. With his laser gun strapped to his side and GPS in hand, he cracks open the hatch and is hit with the scent of warm, humid air that contrasts the cold stale spaceship. Perri pops up from behind him and slaps his shoulders, “Ready when you are, big boy.” Jonathan huffs and gets going.",
-                            "Jonathan starts to acclimate to the smell of mud and algae. It’s fairly dark because of how dense the trees and vines are, with only small patches of light speckling through the leaves. He treads carefully through the shallow water, knowing that at any moment it could become waist deep or he could accidentally step on a sleeping creature.",
-                            "Jonathan pauses and gives Perri a second to catch up to him. “The GPS says we should keep heading this way,” he says, pointing towards an especially shadowed area of the swamp.",
-                            "Perry squints into the darkness with hesitation, “Are you sure? If I were an egg, I would be...I don’t know, anywhere else but there.”",
-                            "Before Jonathan can reassure Perri, a flash of green whips out from the darkness. It strikes Perri to the ground with a wet flop then yanks him out of the water, wrapped around his ankle. The vine-like arm dangles Perri upside down before Jonathan whips out his laser gun to vaporize the vine. Perri worms free from the arm as it slithers back into the depths of the swamp.",
-                            "Helping Perri up, the two start to trudge in the opposite direction to reestablish their footing. Whatever thing that attacked is still nearby, it can be heard sloshing through the water. It can’t be more than 30 feet away but it’s thoroughly camouflaged with the swamp. For all they knew, the entire swamp could be made of the thing.",
-                            "Before they can really catch their breath, the vine attacks once more. But this time it’s multiple vines and slowly the thing starts to reveal itself. A swamp monster, twice their height with numerous green vine arms. A big heavy vine crashes down next to them, sending a swampy wall of water into the air. As it falls, a glimmer catches Jonathan’s eye. Behind this swamp creature, is the pearlescent shining egg that they were sent to pick up.",
-                            "“Perri the egg is behind the monster.”",
-                            "“Oh great.”",
-                            "Jonathan has never been particularly strategic. And in fact, he’s gotten pretty far by pure brute force. So his current plan of attack is to mud wrestle the vine that’s defending the egg. With one vine occupied, Perri takes this opportunity to scamper in the direction of the egg. But the swamp monster, although it doesn’t have any visible eyes, is able to see Perri approach and grabs him with a different vine.",
-                            "“Jonathan, what do I do?!”",
-                            "“Try… talking to it?” Jonathan struggles while still wrestling.",
-                            "“Hi there Mr. Swamp. I really like your… how green you are.” Perri says, giving the vine wrapped around him a hesitant pat, pat. The swamp monster slows, appearing to be listening to Perri’s words. But before Perri can speak again, his phone ringtone goes off.",
-                            "Hey, I just met you. And this is crazy! But-",
-                            "The swamp monster shrieks and Perri shouts, “I don’t think it likes Carly Rae Jepsen!” as the swamp monster waves him around like a doll. With the swamp monster distracted, Jonathan makes his move for the egg. He does a half swimming half crab walking movement to travel through the swamp water as quickly as possible. Perri manages to pick up the phone, “Hi! Yea, we are going to be a little late with the delivery.”",
-                            "Jonathan tucks the egg under one arm and uses the other to shoot more lasers in Perri’s direction. The beams strike the vine and the swamp thing releases Perri into the water below. Jonathan dashes forward, picking up Perri in the process, and makes a run for the ship."
-                        ],
-                        "Some Chicks": [],
-                        "Under the Open Window": [
-                            "The carpet feels grainy under me as I lay in the little patch of sunlight seeping through my open window. A window I pray would shut silently so as to not bring any awareness towards me. A slight breeze blows through, the quiet of the neighborhood falling densely on the day. Eerily, not even the usual sound of birds chirping in the afternoon sound through. I wonder why the world seems so at ease in a moment, immensely dense. Not one day ago was my life turned upside down, the world doing so with it. I slowly sit up to peer out from under the window, to catch any glimpse of life beyond my own. Minutes go by of gazing, my thoughts nonexistent for the first time. Everything has been drained out of me; every possible scenario has gone through my head and yet I still come up empty on the truth of the matter.",
-                            "Twenty-four hours ago to the minute, an unknown siren blared through my city. Then electricity vanished, and cell phone towers failed to reach. The currently silent neighborhood was full of endless noise; panicked conversations from neighbors, cars being packed, doors slamming and traffic filled with raging horns in an attempt to race towards somewhere unknown to me. It was impossible to get a hold of your loved one if not in your immediate surroundings. Amid panic and unsureness, my body halted, fear holding me in my tiny one-bedroom apartment. I knew wherever I tried to go would leave me stuck in the middle of it all with strangers who I did not trust. So I stayed behind and accepted whatever fate rested in my path.",
-                            "Now, I am alone. I move from my spot under the window and crawl over to my remote to check once again if the power has returned. My thumb hovers over the power button, denial holding me back from pressing. When I do, only the expected rush of dread seeps to the pit of my stomach. Not only is my stomach consumed with unease, but hunger as well. It’s been over a day since I’ve had anything to eat. As I consider whether or not I can bear to eat in my anxious state, I hear the sound of footsteps on concrete. I stop dead in my spot, breath coming in short, aggressive spurts due to fear that any essence of existence coming from me would draw this unknown entity near. I hone in on the sound to determine the distance from my open window. The sound echoes on the street outside going at a slow pace; opposite of the steps running through the streets yesterday afternoon.",
-                            "On the floor across the room lies the gleaming pair of scissors, the only source of somewhat protection I could spot in my vicinity. I stare at them, willing them to just conveniently be next to me. The footsteps which I believe to be on the sidewalk directly ahead of my apartment, come to a stop. As tears begin to well in my eyes and threaten to fall, I come up with my action plan. If whatever this is comes near, I will make a run for the scissors, grab the keys to the parking garage, and race to my car and hope there’s an end in sight. My plan seems more probable as I hear the footsteps resume, heading straight for me laying under the open window. My breath quickens, loud spurts of air escape my lungs regardless of my efforts to stop them. The footsteps fall slow, edging closer and closer then come to a complete halt. There’s a certain feeling, when you know that someone or something is near, looking at or for you. My body is ready for flight, because I don’t think I have a fight in me. Silence drags on for what feels like eternity, and my hand resting on the gray carpet trembles, my eyes fixated on them wishing for them to calm. A chill runs down my spine as that feeling of being watched floods over me. Slowly, I lift my head and my stomach sinks.",
-                            "A pair of deep, gray eyes meet mine and I let out a piercing scream. I scramble back from the window, knocking over the cow shaped footrest and crashing into the coffee table. Fear is pulsing through me, keeping me in place, and every plan I had of escape leaves my mind. The gray eyes are in full view, staring through my window but making no move to get in. Racing thoughts entrap my mind, taking in who is in front of me. “Nick?” My neighbor gives no response, just stares blankly at me. “Nick, what’s going on?” Something deep in my bones is sending a warning to me. I’ve known this man for five years. Every morning, he steps outside and waters the complex’s community garden out back. Every morning he says “Good day” excitedly to the neighbors who cross his path. I don’t know what it is, but I do know that this is not that Nick. The man who lovingly pushes his child around the neighborhood in a stroller, letting her run free through the grassy lawn and teaching her how to do cartwheels. The man who waits on the lawn chair out front, drinking a cup of tea while waiting for his wife to get back from her late night nursing shift. This man standing in my window hasn’t an ounce of thought or feeling behind his eyes.",
-                            "“Nick, please tell me what’s going on. That siren, everyone packing up leaving, and then the emptiness and silence…what happened? Why aren’t you gone? Where is your family?” Questions pour out; all the thoughts, the fears, and scenarios I’ve created in my head over the last twenty four hours projecting out in a mess of words. I find my composure as the seconds creep on without a response. Just the same empty stare. “Ok”, I say with a tremble. One last question pushes its way through. “Is it safe outside?”. My eyes dart over every inch of his face; the sunlight shining behind him creating a silhouette of light around his body. All of a sudden, the silhouette breaks as he slowly shakes his head no. My breath catches in my throat as my muscles constrict and seemingly pulls me into myself. Tears well in my eyes instantly, threatening to spill as the realization dawns on me that I am not safe and have no way of knowing why. Without wasting another second, I grab the string of the blinds and pull them down, hiding myself from Nick and the outside world. Tears are now soaking my face as hyperventilation kicks in. My face is tingling from the panic but I manage to speak through it. “Nick, I’m so sorry. I’m scared and…just… if you are ready to speak just knock on the window three times. But I’m going to”, I pause because what am I going to do? I don’t know but all I know is I need to make myself invisible and hide my existence as much as I can in this moment. When I get an expected non response, I frantically begin piling the books on my bookshelf on the floor in a scattered fashion, clearing it so I can place it against the front of the door to attempt a barrier. Once it’s light enough to move, I start placing the books back then grab a chair to place under the doorknob of the back door. I run around the house for the first time, shutting all the blinds in the process and search extensively for anything I could use for protection. Once I stack everything I can, I sit on the couch, unsure of what to do.",
-                            "The sun that was shining over Nick earlier has disappeared and I find myself sitting in the pitch black for the second time since the unknown siren deteriorated the life I knew and understood. The one and only sound of my pounding heart is interrupted by the sound of footfall over the lawn. Is Nick back? Did he ever leave is actually the question. I listen keenly only to realize that it’s not just one person walking this time, but dozens. My body starts shaking uncontrollably once more; the fear is even greater than I ever believed possible. The steps draw towards my closed window like a tide slowly pushing in. One pair of feet particularly lands right outside, about ten feet from where I sit trembling on the couch. That’s when I hear, three slow, eerie knocks."
-                        ],
-                        "Grown Women do as They'd Like": [
-                            "Vandeberg’s mobile home park. Seventy-five homes, twenty-two acres. Half were single wide, and the other half doubles. It was where I hoped to live with Cheryl when we got married—before we’d fought. The police were on the front stoop, so I kicked a few beer cans out of the way and opened the door to see what they wanted.",
-                            "“Hello?” I said.",
-                            "“Cal Owens?” said the one I didn’t know.",
-                            "“That’s him,” said the one I did.",
-                            "“Mac, what are you guys doing here?” I asked.",
-                            "“We’re here just to, well, ask you a few questions, Owns,” he said, “This is Detective Grainger.”",
-                            "“Um, hi, and OK, I guess.”",
-                            "I let them in, a little embarrassed about the mess. I’d been busy with a bender and hadn’t cleaned for days. The police in their clean clothes looked worried about sitting down, but they did anyway. Their heads swiveled, settling briefly on scattered, filthy clothes and the containers of chicken from the gas station.",
-                            "“Cal, Detective Grainger and I want you to tell us what happened after I pulled you over the other day.”",
-                            "“What do you mean, what’d I do after, Mac?”",
-                            "“Yeah, what’d you do after I let you go, exactly.”",
-                            "“I drove out to Cheryl’s place and dropped off some mail. Then I went home, I guess.”",
-                            "Grainger piped up then, “So what happened when you gave her the mail?”",
-                            "“I didn’t see her, she wasn’t home. Nobody was home.”",
-                            "“And you were all by yourself?”",
-                            "“Yep.” Then, turning to Mac, “What is this about?”",
-                            "“Cal, nobody has seen Cheryl.”",
-                            "That sank me. We’d been fighting, and I admit I fucked it up a bit. She left crying. Typical woman, but I could have handled things better. We were supposed to get married, but she wanted to go to school. I shut it down though. She hopped right out of my car in the drive-through. I hadn’t seen her since.",
-                            "“What the fuck do you mean, nobody’s seen her?”",
-                            "“Relax, Mr. Owens, we’re just trying to understand what happened the last time you saw Cheryl”",
-                            "“I don’t know who the fuck you are Grainger, but right now I’m talking to him.” I gestured at Mac and repeated, “What the fuck do you mean, nobody’s seen Cheryl?”",
-                            "“Owns, you were the last one to talk to her best we can tell, the car behind you saw her run off around the corner but that’s it.”",
-                            "“Mr. Owens, we just need you to make a statement at this time,” said Grainger, “We think something may have happened to her, and we’d really like to rule you out.”",
-                            "I blinked at him then. They thought me, that I had, that I’d done something to her!",
-                            "“Mac, what the fuck is this?“",
-                            "“Owns, we're just trying to figure out what happened to her. Maybe she’s just left town, she’s a grown woman, even if she still lives at home.”",
-                            "“I know she’s a grown woman, damnit, we’re getting married!”",
-                            "“Owns, we both know she never agreed to that yet”",
-                            "“Mr. Owens, can you please answer our questions and we’ll be on our way.”",
-                            "“Look, guys, the last I saw her she was bawling her eyes out because I screwed up, I turned my head, and she was gone. Long gone. Then, Mac, you pulled me over on the way to her place. I dropped off a letter and came home. That’s all. That’s the whole show.”",
-                            "“Mac, what’d you pull him over for?” said Grainger",
-                            "“Drinkin n’ dri—” I started.",
-                            "“I pulled him over for littering, Grainger” Mac gave me a hard look.",
-                            "Grainger paused, eyes on Mac for a moment, then nodded once and continued, “What was the letter about, Mr. Owens?”",
-                            "“It was an acceptance letter.”",
-                            "“What kind of acceptance letter?”",
-                            "“A college acceptance letter?”",
-                            "“Why did you have it?”",
-                            "“Because I didn’t want her to go, Detective.”",
-                            "Grainger paused to look at Mac again. Mac just kept looking at me. His eyes were concerned or confused; I don’t know. Maybe pleading.",
-                            "“Who did you give it to when you went to her place?” Grainger continued.",
-                            "“I just shoved it in the box, nobody was home. Then I left.”",
-                            "“Why’d you have her mail?”",
-                            "“I kind of, I didn’t want her to go, so I just, I hid it from her. Nobody, I mean nobody, wants her to leave. Her own mama don’t want her to go! Maybe you ought to go hassle her!”",
-                            "“Mr. Owens, you’re saying you took her letter and hid it from her?”",
-                            "“I mean…”",
-                            "“Relax, we're not talking to you about stealing mail right now.",
-                            "“Yeah, I guess I did,” I admitted.",
-                            "“Have you seen Cheryl since then?”",
-                            "“No.”",
-                            "“Folks have seen you around town drinking pretty heavy the last few days,” said Grainger, “What’s that about?”",
-                            "“You’ve never had a fight with your girl, Detective?”",
-                            "“Not one that she disappears afterwards,” he sneered back at me.",
-                            "Mac shot him a hard look. Then leaned over and whispered something in his ear.",
-                            "“Alright Mac, OK, Mr. Owens, I’m going to leave you my card, we have everything we need for now… Mr. Owens?",
-                            "“Yeah?”",
-                            "“We’ll be seeing you again soon.”",
-                            "As they left I could see Mac throw his hands out, visibly upset with Grainger. Grainger just put his finger in Mac’s face, arguing right back. Then after a tic, Mac put his hands down and they walked back to his car.",
-                            "But they didn’t leave. I watched them through the window, thinking about it all. Cheryl told me she got into it with her mother about leaving town. Cheryl had stormed off when her mom told her she wouldn’t fill out her financial aid forms or help with tuition. She’d come over half drunk, rambling about the FAFSA, and ways around putting her parents on it.",
-                            "“Get married or join the Air Force,” she’d said.",
-                            "Looking past the police car outside I could see the neighborhood kids staring and pointing from the ballfield across the way. And I realized then the whole street was outside, rubbernecking from their door stoops and front lawns. The whole neighborhood is always flipping me shit over how I fast I drive and how loud my music is. Can’t decent people just be left alone? If the cops weren’t still here I’d holler at them.",
-                            "Mac and Grainger talked on the radio for about ten minutes and then Grainger got back out.",
-                            "He banged on the door.",
-                            "“Don’t you have better things—”",
-                            "“Mr. Owens, I think you better come with us. We’d like you to answer some questions down at the station. At this time, you’re being detained for further questioning.",
-                            "Mac was standing at the foot of the stairs, a look of disappointment on his face.",
-                            "“Owns, Cheryl’s mom says she hasn’t gotten any mail for weeks.”",
-                            "“What? I put the letter in the box! I did! Mac, you know I wouldn’t do anything to Cheryl!”",
-                            "“But you did, didn’t you, Mr. Owens.”",
-                            "Then it came to me. “Cheryl’s mom!” I shouted.",
-                            "“Mr. Owens,” Grainger replied skeptically.",
-                            "“No, Cheryl’s Mom, she doesn’t want her to go to school either!”",
-                            "Grainger looked at Mac.",
-                            "“What are you getting at?”",
-                            "“I mean, maybe she was taking her mail too!” They looked at each other and ended up putting me in the back of the car anyway. On the way down the road, they called and asked Cheryl’s mom about it, and she even admitted to taking some of the letters, but she said she hadn’t seen the one I left.",
-                            "That was when the panic faded. I just laughed and laughed. Maybe I deserved to spend the night in jail after all. Cheryl’s Mom had been taking her mail too. I couldn’t help but smile, there in the backseat.",
-                            "“What’s funny?” asked Grainger. Then to Mac, “You think he’s cracking up?”",
-                            "Mac was eyeballing me with concern through the rearview.",
-                            "“Don’t you see? I’m just as bad as Cheryl’s mother.”",
-                            "Grainger hit the brakes, “Something you want to say?”",
-                            "Mac eyed him, then turned around and looked at me.",
-                            "I just couldn’t help but laugh, “I think you guys better call the recruiter.”",
-                            "“What recruiter is that Mr. Owens?”",
-                            "“Air Force,” I giggled.",
-                            "“Owns are you ok?”",
-                            "“He knows more than he’s letting on, Mac,” said Grainger.",
-                            "“Maybe…” said Mac, looking at me, worried.",
-                            "“Look, maybe she lit off out of town to school, maybe she joined the Air Force. She ran out on me in the drive-through, no reason she’d stay with her mom if Cheryl knew she was doing the same things I was.”",
-                            "Mac and Grainger look at me over their shoulders and then back at each other.",
-                            "“She’s a grown woman, she’s going to do what she wants. She doesn’t want to stay here, so why would she?”",
-                            "Cheryl was long gone; I realized nothing bad had happened to her. She’d found the letter and lit out of town. Goodbye me, goodbye mother. I suppose that’s the way it ought to be."
-                        ]
-        };
-
-        const post = posts[i];
-        const paragraphs = content[post.title].map((element, i) => <p key={i}>{element}</p>);
-        
-        return (
-            <article className={styles.detailContainer}>
-                <h1 className={styles.fictionTitle}>{post.title}</h1>
-                <p>Published in <Link href={"../../fiction"}>Fiction</Link></p>;
-                <p className={styles.postDate}>Posted on 2025-06-10</p>
-                <div className={styles.contentContainer}>
-                    {paragraphs}
-                </div>
-            </article>
-        );
-    }
-}
